+++ src/migrator/migrations.sbvr
@@ -65,15 +65,15 @@ Fact Type: migration status is backing off
 Fact Type: migration status has converged time 
 	Necessity: each migration status has at most one converged time
 
-Term: execution time
+Term: execution date
 	Concept Type: Date Time (Type)
 Term: callback
 	Concept Type: Short Text (Type)
 
-Term: scheduled migration
-Fact Type: scheduled migration has migration key
-	Necessity: each scheduled migration has exactly one migration key
-Fact Type: scheduled migration has execution time
-    Necessity: each scheduled migration has exactly one execution time
-Fact Type: scheduled migration has callback
-	Necessity: each scheduled migration has exactly one callback
+Term: scheduled job
+Fact Type: scheduled job has migration key
+	Necessity: each scheduled job has exactly one migration key
+Fact Type: scheduled job has execution date
+    Necessity: each scheduled job has exactly one execution date
+Fact Type: scheduled job has callback
+	Necessity: each scheduled job has exactly one callback
