+++ src/migrator/utils.ts
@@ -108,12 +108,6 @@ export type MigrationStatus = {
 	is_backing_off: boolean;
 };
 
-export type ScheduledMigration = {
-	migration_key: string;
-	execution_time: Date;
-	callback: string;
-};
-
 export const getRunnableAsyncMigrations = (
 	migrations: Migrations,
 ): RunnableAsyncMigrations | undefined => {
