+++ src/server-glue/module.ts
@@ -18,7 +18,7 @@ export * as errors from '../sbvr-api/errors';
 export * as env from '../config-loader/env';
 export * as types from '../sbvr-api/common-types';
 export * as hooks from '../sbvr-api/hooks';
-export * as migrations from '../sbvr-api/migrations';
+export * as jobs from '../sbvr-api/jobs';
 export type { configLoader as ConfigLoader };
 export type { migratorUtils as Migrator };
